fix: satisfy pkg-file-count for skip-permissions test layout
Move skip-permissions policy into pkg/invocations/skippermissions, consolidate service/provider S14 evidence tests, and extract validateWorkerLoadPreflight for pkg-maint.
