+++ docs/internal/processes/invocation-relevant-files.md
@@ -151,16 +151,16 @@ primary-result behavior.
   `RunConfig.InvocationSkipPermissionsOverride`, and forwarded through
   `buildRunServiceConfig` into `service.FactoryServiceConfig` as an ephemeral
   invocation override that must not mutate persisted worker `skipPermissions`.
-  `invocations.EffectiveSkipPermissions` resolves persisted worker config plus
+  `pkg/invocations/skippermissions.EffectiveSkipPermissions` resolves persisted worker config plus
   `FactoryServiceConfig.InvocationSkipPermissionsOverride` when building
   provider-backed worker CLI args in `pkg/service/factory_build.go` and
-  `pkg/runtimehost/build_workers.go`. `invocations.ValidateInvocationSkipPermissionsWorkers`
+  `pkg/runtimehost/build_workers.go`. `skippermissions.ValidateInvocationSkipPermissionsWorkers`
   and `ValidateInvocationSkipPermissionsForWorker` fail closed before worker
   construction when `--skip-permissions` is set but an agent worker uses an
   unsupported CLI provider or local managed model path. S14 regression evidence
-  lives in `pkg/invocations/skip_permissions_evidence_test.go`,
-  `pkg/workers/provider/skip_permissions_evidence_test.go`, and
-  `pkg/service/skip_permissions_evidence_test.go` alongside the story-level
+  lives in `pkg/invocations/skippermissions/skip_permissions_test.go`,
+  `pkg/workers/provider/provider_behavior_test.go`, and
+  `pkg/service/factory_test_helpers_test.go` alongside the story-level
   propagation and fail-closed service tests. `RunConfig.JSONOutput`
   must stay aligned with the shared `InvocationResponse` envelope for both
   successful and non-success invocation results rather than becoming a
