--- pkg/invocations/skip_permissions_evidence_test.go
@@ -1,81 +1,0 @@
-package invocations
-
-import (
-	"testing"
-
-	"github.com/portpowered/infinite-you/pkg/interfaces"
-)
-
-func TestS14SkipPermissionsPrecedenceEvidence(t *testing.T) {
-	t.Parallel()
-
-	overrideTrue := true
-	overrideFalse := false
-
-	tests := []struct {
-		name               string
-		persisted          bool
-		workerType         string
-		invocationOverride *bool
-		want               bool
-	}{
-		{
-			name:       "AbsentOverrideUsesPersistedFalse",
-			persisted:  false,
-			workerType: interfaces.WorkerTypeAgent,
-			want:       false,
-		},
-		{
-			name:               "OverrideTrueWithPersistedFalse",
-			persisted:          false,
-			workerType:         interfaces.WorkerTypeAgent,
-			invocationOverride: &overrideTrue,
-			want:               true,
-		},
-		{
-			name:               "OverrideTrueIgnoredForModelWorker",
-			persisted:          false,
-			workerType:         interfaces.WorkerTypeModel,
-			invocationOverride: &overrideTrue,
-			want:               false,
-		},
-		{
-			name:               "OverrideFalseWithPersistedFalse",
-			persisted:          false,
-			workerType:         interfaces.WorkerTypeAgent,
-			invocationOverride: &overrideFalse,
-			want:               false,
-		},
-		{
-			name:       "PersistedTrueWithoutOverride",
-			persisted:  true,
-			workerType: interfaces.WorkerTypeAgent,
-			want:       true,
-		},
-		{
-			name:               "PersistedTrueWithOverrideFalse",
-			persisted:          true,
-			workerType:         interfaces.WorkerTypeAgent,
-			invocationOverride: &overrideFalse,
-			want:               true,
-		},
-		{
-			name:               "PersistedTrueWithOverrideTrue",
-			persisted:          true,
-			workerType:         interfaces.WorkerTypeAgent,
-			invocationOverride: &overrideTrue,
-			want:               true,
-		},
-	}
-
-	for _, tc := range tests {
-		tc := tc
-		t.Run(tc.name, func(t *testing.T) {
-			t.Parallel()
-			got := EffectiveSkipPermissions(tc.persisted, tc.workerType, tc.invocationOverride)
-			if got != tc.want {
-				t.Fatalf("EffectiveSkipPermissions() = %v, want %v", got, tc.want)
-			}
-		})
-	}
-}
