+++ pkg/invocations/skippermissions/skip_permissions.go
@@ -1,4 +1,4 @@
-package invocations
+package skippermissions
 
 import (
 	"fmt"
